refactor(interfaces): add index files
- update imports
