+++ interfaces/context/IStateContext.ts
