+++ interfaces/data/IAudioData.ts
@@ -2,15 +2,6 @@
  * Defines audio data interfaces and types.
  */
 
-import type Parser from 'rss-parser';
-
-export interface IRssItem extends Parser.Item {
-  [key: string]: any;
-  itunes?: {
-    [key: string]: any;
-  };
-}
-
 /**
  * Audio data interface.
  */
