+++ interfaces/data/index.ts
@@ -1,4 +1,5 @@
 export * from './IAudioData';
 export * from './IEmbedData';
 export * from './IListenData';
+export * from './IRssItem';
 export * from './ISeriesData';
