+++ interfaces/page/IPageProps.ts
@@ -1,4 +1,4 @@
-import { IEmbedConfig } from '@interfaces/embed/IEmbedConfig';
+import { IEmbedConfig } from '@interfaces/embed';
 
 export interface IPageProps {
   config: IEmbedConfig;
