+++ lib/generate/string/generateEmbedUrl.spec.ts
@@ -1,4 +1,4 @@
-import { IEmbedConfig } from '@interfaces/embed/IEmbedConfig';
+import { IEmbedConfig } from '@interfaces/embed';
 import generateEmbedUrl from './generateEmbedUrl';
 
 describe('lib/generate/string', () => {
